Home and Analytics Merged into one page
Removed analytics pages and corresponding code
