+++ app/app/lib/api.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import * as SecureStore from 'expo-secure-store';
 
-const api = axios.create({ baseURL: 'http://172.23.46.114:3000/api' });
+const api = axios.create({ baseURL: 'http://192.168.50.203:3000/api' });
 
 api.interceptors.request.use(async (config) => {
   const token = await SecureStore.getItemAsync('USER_TOKEN');
